requirements.txt
# Core Data Processing
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Environment & Config
python-dotenv>=1.0.0

# Web Scraping & API
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: For OSM data
# osmnx>=1.6.0

# Optional: For Google Places API
# googlemaps>=4.10.0
